requirements.txt
openai-whisper
gradio>=4.0
moviepy==1.0.3
numpy
matplotlib
Pillow
tqdm
